backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "httpx>=0.28.1",
    "python-dotenv>=1.1.0",
    "python-json-logger>=3.3.0",
]
